Compile and link succeeds!

+++ boot/fs.cc
@@ -12,7 +12,7 @@ int boot_close(int file)
     return fs_api.boot_close(file);
 }
 
-int boot_pread(int file, void *buf, size_t bytes, off_t ofs)
+ssize_t boot_pread(int file, void *buf, size_t bytes, off_t ofs)
 {
     return fs_api.boot_pread(file, buf, bytes, ofs);
 }
